Use a dedicated service for managing the MongoClient instances
When using many rivers that connect to the same set of servers it is wasteful to have one MongoClient instance with
a dedicated connection pool for each river.

The MongoClientService is a per-node service, and provides access to MongoClient instances based on the provided list
of server addresses. Note that this means rivers that use the same servers but different options will now get the options
of whichever river definition was processed first. In practice this should not be a problem, as the options are exactly
intended to control resource usage.