Stop implementing Runnable in CollectionSlurper
The CollectionSlurper class is really a helper, and not something that runs in a separate thread. Removing
the Runnable allows us to give a reasonable name to the method that does the initial import, and later on
also allows to provide parameters to the methods rather than to the class.