Start importing from the last timestamp when a river is stopped and started again
The river remembers the timestamp it last imported, but this information wasn't used
when the tailer threads start up. If a river was stopped for a while (for example because
it failed and was restarted) data between the last import time and the current shard
oplog time was never processed.

Note that the test does some aggressive reflection to restart the river, this would become
obsolete with #407 merged.