Keep close track of the processed oplog entries
When scanning the oplog there are many entries that are not directly relevant
for the current river: administrative messages, no-op entries, entries for other collections, ... .
Previously we simply ignored these.

This "ignoring" does have quite grave consequences:
1. When the river is restarted it may have to re-process many of them, especially
   when it is a for a collection with little activity, in a database with very
   active other collections.
2. As a direct consequence of that the amount of oplog entries
   that need to be available must be such that of *every* collection the newest
   eligible entry is still in the collection.

Especially the last point is problematic when the cluster topology itself is very
dynamic, as it can happen in automatically managed clusters in a cloud environment.

This commit changes the approach so that for each bulk of oplog entries at least
the timestamp in the river state is updated: either because there was at least one
eligible entry, or, failing that, by introducing a "no-op" request into the index stream.